User and Post Repositories

+++ backend/src/main/java/com/orchard/service/AccountService.java
@@ -21,7 +21,7 @@ public interface AccountService {
 	
 	public void updateUser(User user);
 	
-	public User findById(Long id);
+	public User findById(Integer id);
 	
 	public void deleteUser(User user);
