+++ backend/src/main/java/com/orchard/service/PostService.java
@@ -14,7 +14,7 @@ public interface PostService {
 	
 	public List<Post> postList();
 	
-	public Post getPostById(Long id);
+	public Post getPostById(Integer id);
 	
 	public List<Post> findPostByUsername(String username);
